[zeta-chess] Releasing all GPU memory.
